@@ -1,12 +1,13 @@
 package dev.gtuk.diga.dtos
 
-import java.util.Date
-import javax.validation.constraints.NotEmpty
-import javax.validation.constraints.Size
+import jakarta.validation.constraints.NotEmpty
+import jakarta.validation.constraints.Size
+import java.util.*
+
 
 data class BillingRequest(
-    @get:NotEmpty @get:Size(min = 16, max = 16) val code: String,
-    @get:NotEmpty @get:Size(min = 8, max = 8) val digavId: String,
-    val dayOfServiceProvision: Date,
-    @get:NotEmpty val invoiceNumber: String
+        @get:NotEmpty @get:Size(min = 16, max = 16) val code: String,
+        @get:NotEmpty @get:Size(min = 8, max = 8) val digavId: String,
+        val dayOfServiceProvision: Date,
+        @get:NotEmpty val invoiceNumber: String
 )
